requirements/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<2.0.0
psycopg2>=2.9.0,<3.0.0
alembic>=1.6.0,<2.0.0
python-dotenv>=0.19.0,<1.0.0
pytest>=6.2.4,<7.1.3
pytest-mock==3.10.0
requests>=2.25.1,<3.0.0
